Tests/LibWeb/Text/expected/WebAudio/OfflineAudioContext-AudioWorklet-process-return-false.txt

sample0: 1.0
sample128: 0.0
